requirements.txt
tqdm==4.28.1
Pillow==6.0.0
coloredlogs==10.0
